Crate mdman

mdman markdown to man converter.

Modules

format 🔒

hbs 🔒

Handlebars template processing.

util 🔒

Enums

Format

The output formats supported by mdman.

Functions

convert

Converts the handlebars markdown file at the given path into the given format, returning the translated result.

extract_section

join_url 🔒

md_parser 🔒

Creates a new markdown parser with the given input.

Type Aliases

EventIter 🔒

Pulldown-cmark iterator yielding an (event, range) tuple.

ManMap

Mapping of (name, section) of a man page to a URL.

Section

A man section.