Module mdman::hbs

Handlebars template processing.

Structs

ManLinkHelper 🔒

Helper for {{man name section}} expression.

OptionHelper 🔒

Helper for {{#option}} block.

OptionsHelper 🔒

Helper for {{#options}} block.

lower

Functions

expand

Processes the handlebars template at the given file.

in_options 🔒

Whether or not the context is currently inside a {{#options}} block.

remove_from_context 🔒

Removes a variable from the context.

set_decorator 🔒

{{*set var=value}} decorator.

set_in_context 🔒

Sets a variable to a value within the context.

Type Aliases

FormatterRef 🔒